+++ include/bulk/messages.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstring>
 #include <memory>
 #include <tuple>
 #include <vector>
@@ -34,6 +35,16 @@ struct message_t<T, typename std::enable_if_t<(sizeof...(Ts) == 0)>, Ts...> {
     T content;
 };
 
+template <typename T>
+struct message_t<T[], void> {
+    std::vector<T> content;
+};
+
+template <typename T, typename... Ts>
+struct message_t<T[], typename std::enable_if_t<(sizeof...(Ts) > 0)>, Ts...> {
+    std::tuple<std::vector<T>, Ts...> content;
+};
+
 template <typename T, typename... Ts>
 struct message : public message_t<T, void, Ts...> {};
 
@@ -53,6 +64,8 @@ class queue_base {
     virtual void clear_() = 0;
 
     virtual void unsafe_push_back(void* msg) = 0;
+    virtual void unsafe_push_array(int count, size_t size, void* msg,
+                                   size_t size_of_other, void* other) = 0;
 };
 
 /**
@@ -66,6 +79,10 @@ class queue {
   public:
     using message_type = decltype(message<T, Ts...>::content);
     using iterator = typename std::vector<message_type>::iterator;
+
+    // If T = E[] is an array type, then element_type is E, otherwise it is T
+    using element_type =
+        typename std::remove_pointer<typename std::decay<T>::type>::type;
 
     /**
      * A queue is a mailbox for messages of a given type.
@@ -88,6 +105,26 @@ class queue {
         void send(Us... args) {
             message_type msg = {args...};
             q_.impl_->send_(t_, msg);
+        }
+
+        /* Enabled if first content type is an array. */
+        // We have to let this depend on its own template parameter to allow
+        // SFINAE to kick in
+        template <typename U = T,
+                  typename = typename std::enable_if_t<
+                      std::is_array<U>::value && (sizeof...(Ts) == 0)>>
+        void send_many(std::vector<element_type> msgs) {
+            q_.impl_->send_many_(t_, msgs.size(), msgs.data(), nullptr, 0);
+        }
+
+        template <typename U = T,
+                  typename = typename std::enable_if_t<
+                      std::is_array<U>::value && (sizeof...(Ts) > 0)>>
+        void send_many(std::vector<element_type> msgs, Ts... args,
+                       void* = nullptr) {
+            message_type msg = {std::vector<element_type>{}, args...};
+            q_.impl_->send_many_(t_, msgs.size(), msgs.data(), &msg,
+                                 sizeof(message_type));
         }
 
       private:
@@ -170,6 +207,12 @@ class queue {
             world_.send_(t, id_, &m, sizeof(m));
         }
 
+        void send_many_(int t, int count, element_type* m, void* other,
+                        size_t size_of_other) {
+            world_.send_many_(t, id_, m, sizeof(element_type), count, other,
+                              size_of_other);
+        }
+
         void* get_buffer_(int size_in_bytes) override {
             data_.resize(size_in_bytes / sizeof(message_type));
             return &data_[0];
@@ -177,6 +220,24 @@ class queue {
 
         void unsafe_push_back(void* msg) override {
             data_.push_back(*static_cast<message_type*>(msg));
+        }
+
+        void unsafe_push_array(int count, size_t size, void* data,
+                               size_t size_of_other, void* other) override {
+            if (size_of_other == 0) {
+                data_.push_back(message_type{});
+                auto& msg =
+                    *((std::vector<element_type>*)(&(data_[data_.size() - 1])));
+                msg.resize(count);
+                memcpy(msg.data(), data, size * count);
+            } else {
+                data_.push_back(*static_cast<message_type*>(other));
+                using real_type = std::tuple<std::vector<element_type>, Ts...>;
+                auto& msg = *((std::vector<element_type>*)(&(
+                    std::get<0>(*(real_type*)&data_[data_.size() - 1]))));
+                msg.resize(count);
+                memcpy(msg.data(), data, size * count);
+            }
         }
 
         void clear_() override { data_.clear(); }