+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 ### Added
 
 - Add a 'send many' primitive
+- Add support for array components in messages, i.e. `queue<T[], U, V, ...>`
 
 ### Fixed